style changes

+++ explore.py
@@ -5,8 +5,8 @@ import matplotlib.pyplot as plt
 df = pd.read_csv('./datasets/apartment-prices.csv')
 df['datetime'] = pd.to_datetime(df['sale_date'])
 
-print(df.groupby('rooms').price_per_area.mean())
-print(df.groupby('rooms').living_space.mean())
+print(df.groupby('rooms').price_per_area.mean().astype(int))
+print(df.groupby('rooms').living_space.mean().astype(int))
 print(df.groupby('street_id').price_per_area.mean().astype(int).sort_values(ascending=False).head())
 
 # print(df.rooms.value_counts())
